+++ data_safe_haven/commands/deploy_sre.py
@@ -5,8 +5,8 @@ from data_safe_haven.exceptions import (
 )
 from data_safe_haven.external import GraphApi
 from data_safe_haven.functions import alphanumeric, bcrypt_salt, password
+from data_safe_haven.infrastructure import PulumiSHMStack, PulumiSREStack
 from data_safe_haven.provisioning import SREProvisioningManager
-from data_safe_haven.pulumi import PulumiSHMStack, PulumiSREStack
 from data_safe_haven.utility import DatabaseSystem, SoftwarePackageCategory
