+++ data_safe_haven/infrastructure/components/shm_bastion.py
