+++ data_safe_haven/infrastructure/components/shm_domain_controllers.py
@@ -5,8 +5,8 @@ from collections.abc import Sequence
 from pulumi import ComponentResource, Input, Output, ResourceOptions
 from pulumi_azure_native import network, resources
 
-from data_safe_haven.pulumi.common import get_name_from_subnet
-from data_safe_haven.pulumi.dynamic.remote_powershell import (
+from data_safe_haven.infrastructure.common import get_name_from_subnet
+from data_safe_haven.infrastructure.dynamic.remote_powershell import (
     RemoteScript,
     RemoteScriptProps,
 )
