+++ data_safe_haven/infrastructure/components/shm_firewall.py
@@ -2,7 +2,7 @@
 from pulumi import ComponentResource, Input, Output, ResourceOptions
 from pulumi_azure_native import network
 
-from data_safe_haven.pulumi.common import SREIpRanges, get_id_from_subnet
+from data_safe_haven.infrastructure.common import SREIpRanges, get_id_from_subnet
 
 
 class SHMFirewallProps:
