+++ data_safe_haven/infrastructure/components/shm_monitoring.py
@@ -14,7 +14,7 @@ from data_safe_haven.functions import (
     replace_separators,
     time_as_string,
 )
-from data_safe_haven.pulumi.common import get_id_from_subnet
+from data_safe_haven.infrastructure.common import get_id_from_subnet
 
 
 class SHMMonitoringProps:
