+++ data_safe_haven/infrastructure/components/sre_networking.py
@@ -3,7 +3,7 @@ from pulumi import ComponentResource, Input, Output, ResourceOptions
 from pulumi_azure_native import network, resources
 
 from data_safe_haven.functions import alphanumeric, ordered_private_dns_zones
-from data_safe_haven.pulumi.common import (
+from data_safe_haven.infrastructure.common import (
     NetworkingPriorities,
     SREDnsIpRanges,
     SREIpRanges,
