+++ data_safe_haven/infrastructure/components/sre_software_repositories.py
@@ -4,10 +4,10 @@ import pathlib
 from pulumi import ComponentResource, Input, Output, ResourceOptions
 from pulumi_azure_native import containerinstance, network, storage
 
-from data_safe_haven.pulumi.common import (
+from data_safe_haven.infrastructure.common import (
     get_ip_address_from_container_group,
 )
-from data_safe_haven.pulumi.dynamic.file_share_file import (
+from data_safe_haven.infrastructure.dynamic.file_share_file import (
     FileShareFile,
     FileShareFileProps,
 )
