+++ data_safe_haven/infrastructure/components/virtual_machine.py
