+++ data_safe_haven/infrastructure/declarative_shm.py
