+++ data_safe_haven/infrastructure/declarative_sre.py
