+++ data_safe_haven/infrastructure/dynamic/azuread_application.py
