+++ data_safe_haven/infrastructure/dynamic/compiled_dsc.py
