+++ data_safe_haven/infrastructure/dynamic/dsh_resource_provider.py
