+++ data_safe_haven/infrastructure/dynamic/file_share_file.py
