+++ data_safe_haven/infrastructure/dynamic/ssl_certificate.py
