+++ data_safe_haven/infrastructure/pulumi_stack.py
