+++ data_safe_haven/provisioning/shm_provisioning_manager.py
@@ -1,6 +1,6 @@
 """Provisioning manager for a deployed SHM."""
 from data_safe_haven.external import AzureApi
-from data_safe_haven.pulumi import PulumiSHMStack
+from data_safe_haven.infrastructure import PulumiSHMStack
 
 
 class SHMProvisioningManager:
